AlterQuery supports AND now

+++ webapp-factory/src/main/java/sbcm/supplier/RunBenchmark.java
@@ -114,12 +114,12 @@ public class RunBenchmark extends Role {
 
 	@Override
 	protected void doAction() {
-		woodenstaff = new Integer(500);
-		igniter = new Integer(500);
-		propellant = new Integer(500);
-		effectLoadRed = new Integer(500);
-		effectLoadBlue = new Integer(500);
-		effectLoadGreen = new Integer(500);
+		woodenstaff = new Integer(1500);
+		igniter = new Integer(1500);
+		propellant = new Integer(1500);
+		effectLoadRed = new Integer(1500);
+		effectLoadBlue = new Integer(1500);
+		effectLoadGreen = new Integer(1500);
 		errorRate = new Integer(5);
 		deliver();
 	}
